requirements.txt
discord
discord.py
dislash.py